# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Sumé LMS API for Course Microservice",
        "title": "Sumé LMS Course API",
        "termsOfService": "https://sumelms.com/docs/terms",
        "contact": {
            "name": "LMS Support",
            "url": "https://sumelms.com/support",
            "email": "support@sumelms.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/matrices/{matrix_uuid}/subjects": {
            "get": {
                "description": "List matrix subjects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix_subjects"
                ],
                "summary": "List matrix subjects",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Matrix UUID",
                        "name": "matrix_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.ListMatrixSubjectsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/matrices/{matrix_uuid}/subjects/": {
            "post": {
                "description": "Create a new matrix subject",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix_subjects"
                ],
                "summary": "Create matrix subject",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Matrix UUID",
                        "name": "matrix_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add Matrix Subject",
                        "name": "matrix_subject",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoints.CreateMatrixSubjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.CreateMatrixSubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/matrices/{matrix_uuid}/subjects/{subject_uuid}": {
            "get": {
                "description": "Find a new matrix subject",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix_subjects"
                ],
                "summary": "Find matrix subject",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Matrix UUID",
                        "name": "matrix_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subject UUID",
                        "name": "subject_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.FindMatrixSubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update a matrix subject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix_subjects"
                ],
                "summary": "Update matrix subject",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Matrix UUID",
                        "name": "matrix_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subject UUID",
                        "name": "subject_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update MatrixSubject",
                        "name": "matrix_subject",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoints.UpdateMatrixSubjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.UpdateMatrixSubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete a new matrix subject",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix_subjects"
                ],
                "summary": "Delete matrix subject",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Matrix UUID",
                        "name": "matrix_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subject UUID",
                        "name": "subject_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeletedMatrixSubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/subjects/{uuid}": {
            "get": {
                "description": "Find a new subject",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subjects"
                ],
                "summary": "Find subject",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subject UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.FindSubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update a subject",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subjects"
                ],
                "summary": "Update subject",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subject UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Subject",
                        "name": "subject",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoints.UpdateSubjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.UpdateSubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete a new subject",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subjects"
                ],
                "summary": "Delete subject",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Subject UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeleteSubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "endpoints.CreateMatrixSubjectRequest": {
            "type": "object",
            "required": [
                "group",
                "matrix_uuid",
                "subject_uuid"
            ],
            "properties": {
                "group": {
                    "type": "string"
                },
                "is_required": {
                    "type": "boolean"
                },
                "matrix_uuid": {
                    "type": "string"
                },
                "subject_uuid": {
                    "type": "string"
                }
            }
        },
        "endpoints.CreateMatrixSubjectResponse": {
            "type": "object",
            "properties": {
                "matrix_subject": {
                    "$ref": "#/definitions/endpoints.MatrixSubjectResponse"
                }
            }
        },
...
        "endpoints.DeleteSubjectResponse": {
            "type": "object",
            "properties": {
                "subject": {
                    "$ref": "#/definitions/endpoints.DeletedSubjectResponse"
                }
            }
        },
...
        "endpoints.DeletedMatrixSubjectResponse": {
            "type": "object",
            "properties": {
                "deleted_at": {
                    "type": "string"
                },
                "matrix_uuid": {
                    "type": "string"
                },
                "subject_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "endpoints.DeletedSubjectResponse": {
            "type": "object",
            "properties": {
                "deleted_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "endpoints.FindMatrixSubjectResponse": {
            "type": "object",
            "properties": {
                "matrix_subject": {
                    "$ref": "#/definitions/endpoints.MatrixSubjectResponse"
                }
            }
        },
        "endpoints.FindSubjectResponse": {
            "type": "object",
            "properties": {
                "subject": {
                    "$ref": "#/definitions/endpoints.SubjectResponse"
                }
            }
        },
...
        "endpoints.ListMatrixSubjectsResponse": {
            "type": "object",
            "properties": {
                "matrix_subjects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/endpoints.MatrixSubjectResponse"
                    }
                }
            }
        },
...
        "endpoints.MatrixSubjectResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "is_required": {
                    "type": "boolean"
                },
                "matrix_uuid": {
                    "type": "string"
                },
                "subject_uuid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "endpoints.SubjectResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "credit": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "objective": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "workload": {
                    "type": "number"
                }
            }
        },
...
        "endpoints.UpdateMatrixSubjectRequest": {
            "type": "object",
            "required": [
                "group",
                "matrix_uuid",
                "subject_uuid"
            ],
            "properties": {
                "group": {
                    "type": "string"
                },
                "is_required": {
                    "type": "boolean"
                },
                "matrix_uuid": {
                    "type": "string"
                },
                "subject_uuid": {
                    "type": "string"
                }
            }
        },
        "endpoints.UpdateMatrixSubjectResponse": {
            "type": "object",
            "properties": {
                "matrix_subject": {
                    "$ref": "#/definitions/endpoints.MatrixSubjectResponse"
                }
            }
        },
        "endpoints.UpdateSubjectRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "uuid"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 45
                },
                "credit": {
                    "type": "number"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "objective": {
                    "type": "string",
                    "maxLength": 245
                },
                "published_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "workload": {
                    "type": "number"
                }
            }
        },
        "endpoints.UpdateSubjectResponse": {
            "type": "object",
            "properties": {
                "subject": {
                    "$ref": "#/definitions/endpoints.SubjectResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://sso.sumelms.com/oauth/authorize",
            "tokenUrl": "https://sso.sumelms.com/oauth/token",
            "scopes": {
                "admin": "Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://sso.sumelms.com/oauth/token",
            "scopes": {
                "admin": "Grants read and write access to administrative information",
                "write": "Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://sso.sumelms.com/oauth/authorize",
            "scopes": {
                "admin": "Grants read and write access to administrative information",
                "write": "Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://sso.sumelms.com/oauth/token",
...
}